bug(appfix): fix errors in application

+++ client/src/components/sections/ManageUser.jsx
@@ -1,6 +1,8 @@
 import React, { PropTypes } from 'react';
+import swal from 'sweetalert';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
+import sweetAlert from '../../alert.js';
 import UserList from './UserList.jsx';
 import Paginate from './Paginate.jsx';
 import * as userActions from '../../actions/userAction';
@@ -68,7 +70,12 @@ class ManageUser extends React.Component {
    * @memberof ManageUser
    */
   handleClick(userId) {//eslint-disable-line
-    this.props.actions.deleteAcc(userId);
+    const text = 'You will not be able to recover this user';
+    swal(sweetAlert(text),
+      () => {
+        this.props.actions.deleteAcc(userId);
+        swal('Deleted!', 'The user has been deleted.', 'success');
+      });
   }
   /**
  *
@@ -124,7 +131,10 @@ class ManageUser extends React.Component {
                 <SearchBox onChange={this.handleSearch} />
               </div>
             </div>
-            <UserList users={users} onClick={this.handleClick} />
+            <div>
+              {(this.state.searching && users.length === 0) ?
+                <center> <p><i className="material-icons">warning</i> User Not found </p> </center> : <UserList users={users} onClick={this.handleClick} />}
+            </div>
             {!this.state.searching ?
               <div className="col s12">
                 <Paginate
@@ -134,7 +144,7 @@ class ManageUser extends React.Component {
                 />
               </div> : ''}
           </div>
-          : <img src="default.gif" alt="" />
+          : <img src="default.gif" alt="Not found" />
         }
       </div>
     );
